# kvgraph: optimize iterSources to have logarithmic complexity