--- a/plugins/kvscheduler/internal/graph/edge_lookup.go
+++ b/plugins/kvscheduler/internal/graph/edge_lookup.go
@@ -15,13 +15,18 @@
 package graph
 
 import (
+	"fmt"
+	"reflect"
 	"sort"
 	"strings"
 )
 
 const (
-	initNodeKeysCap = 1000
-	initEdgesCap    = 10000
+	initNodeKeysCap       = 1000
+	initEdgesCap          = 10000
+	initDirDepthBoundsCap = 10
+
+	dirSeparator = "/"
 )
 
 // edgeLookup is a helper tool used internally by kvgraph for **efficient** lookups
@@ -33,6 +38,10 @@ type edgeLookup struct {
 
 	edges        []edge
 	removedEdges int
+	// edges are first sorted (split) by the number of target key components (directories)
+	// -> dirDepthBounds[dirCount] = index of the first edge in <edges> whose
+	//    targetKey consists of <dirCount> directories (incl. the last suffix)
+	dirDepthBounds []int
 
 	overlay  *edgeLookup
 	underlay *edgeLookup
@@ -57,9 +66,10 @@ type edge struct {
 
 func newEdgeLookup() *edgeLookup {
 	return &edgeLookup{
-		nodeKeys:    make([]nodeKey, 0, initNodeKeysCap),
-		nodeKeysMap: make(map[string]bool),
-		edges:       make([]edge, 0, initEdgesCap),
+		nodeKeys:       make([]nodeKey, 0, initNodeKeysCap),
+		nodeKeysMap:    make(map[string]bool),
+		edges:          make([]edge, 0, initEdgesCap),
+		dirDepthBounds: make([]int, 0, initDirDepthBoundsCap),
 	}
 }
 
@@ -68,6 +78,7 @@ func (el *edgeLookup) reset() {
 	el.nodeKeys = el.nodeKeys[:0]
 	el.removedNodes = 0
 	el.edges = el.edges[:0]
+	el.dirDepthBounds = el.dirDepthBounds[:0]
 	el.removedEdges = 0
 }
 
@@ -75,16 +86,19 @@ func (el *edgeLookup) makeOverlay() *edgeLookup {
 	if el.overlay == nil {
 		// create overlay for the first time
 		el.overlay = &edgeLookup{
-			nodeKeys: make([]nodeKey, 0, max(len(el.nodeKeys), initNodeKeysCap)),
-			edges:    make([]edge, 0, max(len(el.edges), initEdgesCap)),
-			underlay: el,
+			nodeKeys:       make([]nodeKey, 0, max(len(el.nodeKeys), initNodeKeysCap)),
+			edges:          make([]edge, 0, max(len(el.edges), initEdgesCap)),
+			dirDepthBounds: make([]int, 0, max(len(el.dirDepthBounds), initDirDepthBoundsCap)),
+			underlay:       el,
 		}
 	}
 	// re-use previously allocated memory
 	el.overlay.resizeNodeKeys(len(el.nodeKeys))
 	el.overlay.resizeEdges(len(el.edges))
+	el.overlay.resizeDirDepthBounds(len(el.dirDepthBounds))
 	copy(el.overlay.nodeKeys, el.nodeKeys)
 	copy(el.overlay.edges, el.edges)
+	copy(el.overlay.dirDepthBounds, el.dirDepthBounds)
 	el.overlay.nodeKeysMap = make(map[string]bool)
 	el.overlay.removedEdges = el.removedEdges
 	el.overlay.removedNodes = el.removedNodes
@@ -107,8 +121,10 @@ func (el *edgeLookup) saveOverlay() {
 	el.nodeKeysMap = make(map[string]bool) // clear
 	el.underlay.resizeNodeKeys(len(el.nodeKeys))
 	el.underlay.resizeEdges(len(el.edges))
+	el.underlay.resizeDirDepthBounds(len(el.dirDepthBounds))
 	copy(el.underlay.nodeKeys, el.nodeKeys)
 	copy(el.underlay.edges, el.edges)
+	copy(el.underlay.dirDepthBounds, el.dirDepthBounds)
 }
 
 // O(log(n))
@@ -159,7 +175,9 @@ func (el *edgeLookup) nodeKeyIdx(key string) int {
 
 // O(log(m))
 func (el *edgeLookup) addEdge(e edge) {
-	idx := el.edgeIdx(e)
+	e.targetKey = trimTrailingDirSep(e.targetKey)
+	dirDepth := getDirDepth(e.targetKey)
+	idx := el.edgeIdx(e, dirDepth)
 	if idx < len(el.edges) {
 		equal, _ := e.compare(el.edges[idx])
 		if equal {
@@ -176,11 +194,19 @@ func (el *edgeLookup) addEdge(e edge) {
 	}
 	el.edges[idx] = e
 	el.edges[idx].removed = false
+	for i := dirDepth + 1; i < len(el.dirDepthBounds); i++ {
+		el.dirDepthBounds[i]++
+	}
+	for i := len(el.dirDepthBounds); i <= dirDepth; i++ {
+		el.dirDepthBounds = append(el.dirDepthBounds, len(el.edges)-1)
+	}
 }
 
 // O(log(m)) amortized
 func (el *edgeLookup) delEdge(e edge) {
-	idx := el.edgeIdx(e)
+	e.targetKey = trimTrailingDirSep(e.targetKey)
+	dirDepth := getDirDepth(e.targetKey)
+	idx := el.edgeIdx(e, dirDepth)
 	if idx <= len(el.edges) {
 		equal, _ := e.compare(el.edges[idx])
 		if equal && !el.edges[idx].removed {
@@ -194,12 +220,30 @@ func (el *edgeLookup) delEdge(e edge) {
 }
 
 // O(log(m))
-func (el *edgeLookup) edgeIdx(e edge) int {
-	return sort.Search(len(el.edges),
+func (el *edgeLookup) edgeIdx(e edge, dirDepth int) int {
+	begin, end := el.getDirDepthBounds(dirDepth)
+	if begin == end {
+		return begin
+	}
+	return begin + sort.Search(end-begin,
 		func(i int) bool {
-			_, order := e.compare(el.edges[i])
+			_, order := e.compare(el.edges[begin+i])
 			return order <= 0
 		})
+}
+
+func (el *edgeLookup) getDirDepthBounds(dirDepth int) (begin, end int) {
+	if dirDepth < len(el.dirDepthBounds) {
+		begin = el.dirDepthBounds[dirDepth]
+	} else {
+		begin = len(el.edges)
+	}
+	if dirDepth < len(el.dirDepthBounds)-1 {
+		end = el.dirDepthBounds[dirDepth+1]
+	} else {
+		end = len(el.edges)
+	}
+	return
 }
 
 // for prefix: O(log(n)) (assuming O(1) matched keys)
@@ -244,46 +288,28 @@ func (el *edgeLookup) iterTargets(key string, isPrefix bool, cb func(targetNode 
 
 // O(log(m)) (assuming O(1) matched sources)
 func (el *edgeLookup) iterSources(targetKey string, cb func(sourceNode, relation, label string)) {
-	// iter over key selectors without key prefixes
-	for i := range el.edges {
-		if el.edges[i].targetKey != "" {
-			break
+	targetKey = trimTrailingDirSep(targetKey)
+
+	var dirDepth int
+	for i := 0; i <= len(targetKey); i++ {
+		prefix := i < len(targetKey)
+		if i == 0 || !prefix || targetKey[i] == dirSeparator[0] {
+			idx := el.edgeIdx(edge{targetKey: targetKey[:i]}, dirDepth)
+			_, end := el.getDirDepthBounds(dirDepth)
+			for j := idx; j < end; j++ {
+				if el.edges[j].targetKey != targetKey[:i] {
+					break
+				}
+				if prefix && !el.edges[j].isPrefix {
+					continue
+				}
+				if el.edges[j].removed {
+					continue
+				}
+				cb(el.edges[j].sourceNode, el.edges[j].relation, el.edges[j].label)
+			}
+			dirDepth++
 		}
-		if el.edges[i].removed {
-			continue
-		}
-		if el.edges[i].isPrefix {
-			cb(el.edges[i].sourceNode, el.edges[i].relation, el.edges[i].label)
-		}
-	}
-	idx := el.edgeIdx(edge{targetKey: targetKey})
-	// iter over key prefixes (smaller than the length of targetKey)
-	for i := idx - 1; i >= 0; i-- {
-		if el.edges[i].targetKey == "" {
-			// empty key prefixes already iterated above
-			break
-		}
-		if el.edges[i].removed {
-			continue
-		}
-		if !el.edges[i].isPrefix {
-			continue
-		}
-		if !strings.HasPrefix(targetKey, el.edges[i].targetKey) {
-			//break
-			continue // TODO: damn...
-		}
-		cb(el.edges[i].sourceNode, el.edges[i].relation, el.edges[i].label)
-	}
-	// iter over exactly matching key(-prefixes)
-	for i := idx; i < len(el.edges); i++ {
-		if el.edges[i].targetKey != targetKey {
-			break
-		}
-		if el.edges[i].removed {
-			continue
-		}
-		cb(el.edges[i].sourceNode, el.edges[i].relation, el.edges[i].label)
 	}
 }
 
@@ -305,16 +331,33 @@ func (el *edgeLookup) gcNodeKeys() {
 // O(m)
 func (el *edgeLookup) gcEdges() {
 	var next int
-	for i := range el.edges {
-		if !el.edges[i].removed {
-			if next < i {
-				el.edges[next] = el.edges[i]
-			}
-			next++
+	for dIdx, curBound := range el.dirDepthBounds {
+		newBound := next
+		nextBound := len(el.edges)
+		if dIdx < len(el.dirDepthBounds)-1 {
+			nextBound = el.dirDepthBounds[dIdx+1]
 		}
+		for i := curBound; i < nextBound; i++ {
+			if !el.edges[i].removed {
+				if next < i {
+					el.edges[next] = el.edges[i]
+				}
+				next++
+			}
+		}
+		el.dirDepthBounds[dIdx] = newBound
 	}
+
 	el.edges = el.edges[:next]
 	el.removedEdges = 0
+
+	dIdx := len(el.dirDepthBounds) - 1
+	for ; dIdx >= 0; dIdx-- {
+		if el.dirDepthBounds[dIdx] < len(el.edges) {
+			break
+		}
+	}
+	el.dirDepthBounds = el.dirDepthBounds[:dIdx+1]
 }
 
 func (el *edgeLookup) resizeNodeKeys(size int) {
@@ -329,6 +372,44 @@ func (el *edgeLookup) resizeEdges(size int) {
 		el.edges = make([]edge, size)
 	}
 	el.edges = el.edges[0:size]
+}
+
+func (el *edgeLookup) resizeDirDepthBounds(size int) {
+	if cap(el.dirDepthBounds) < size {
+		el.dirDepthBounds = make([]int, size)
+	}
+	el.dirDepthBounds = el.dirDepthBounds[0:size]
+}
+
+// for UTs
+func (el *edgeLookup) verifyDirDepthBounds() error {
+	expBounds := []int{}
+	dirDepth := -1
+	for i := range el.edges {
+		tk := el.edges[i].targetKey
+		if len(tk) > 0 && tk[len(tk)-1] == dirSeparator[0] {
+			return fmt.Errorf("edge with targetKey ending with dir separator: %s", tk)
+		}
+		var tkDirDepth int
+		if tk != "" {
+			tkDirDepth = len(strings.Split(tk, dirSeparator))
+		}
+
+		if tkDirDepth < dirDepth {
+			return fmt.Errorf("edge with targetKey inserted at a wrong dir depth (%d): %s",
+				dirDepth, tk)
+		}
+		for j := dirDepth + 1; j <= tkDirDepth; j++ {
+			expBounds = append(expBounds, i)
+		}
+		dirDepth = tkDirDepth
+	}
+	// bad performance of this is OK, the method is used only in unit tests
+	if !reflect.DeepEqual(el.dirDepthBounds, expBounds) {
+		return fmt.Errorf("unexpected dir-depth bounds: expected=%v, actual=%v (edges=%+v)",
+			expBounds, el.dirDepthBounds, el.edges)
+	}
+	return nil
 }
 
 func (e edge) compare(e2 edge) (equal bool, order int) {
@@ -370,4 +451,23 @@ func max(a, b int) int {
 		return a
 	}
 	return b
+}
+
+func trimTrailingDirSep(s string) string {
+	for len(s) > 0 && s[0] == dirSeparator[0] {
+		s = s[1:]
+	}
+	for len(s) > 0 && s[len(s)-1] == dirSeparator[0] {
+		s = s[:len(s)-1]
+	}
+	return s
+}
+
+func getDirDepth(s string) int {
+	var depth int
+	if len(s) > 0 {
+		depth++ // include last suffix (assuming no trailing separator)
+	}
+	depth += strings.Count(s, dirSeparator)
+	return depth
 }