Graph benchmark shows that it is better now for higher number of graph nodes/edges.
For example, before:
```
nchmarkScaleWithoutRecInPlace/1-2           	   50000	     27.311 ns/op
BenchmarkScaleWithoutRecInPlace/10-2          	   10000	    236.857 ns/op
BenchmarkScaleWithoutRecInPlace/100-2         	     300	   4.276.406 ns/op
BenchmarkScaleWithoutRecInPlace/1000-2        	      20	  76.376.488 ns/op
BenchmarkScaleWithoutRecInPlace/10000-2       	       1	3.530.690.159 ns/op
```
Now:
```
BenchmarkScaleWithoutRecInPlace/1-2           	   50000	     27.593 ns/op
BenchmarkScaleWithoutRecInPlace/10-2          	   10000	    246.341 ns/op
BenchmarkScaleWithoutRecInPlace/100-2         	     300	   4.484.061 ns/op
BenchmarkScaleWithoutRecInPlace/1000-2        	      20	  56.016.241 ns/op
BenchmarkScaleWithoutRecInPlace/10000-2       	       2	 936.665.104 ns/op
```
There are some extra calculations now that add some O(1) complexity, worsening the small-scale scenarios, but overall improving scalability.

iterSources is no longer in top10.
Before:
```
(pprof) top
Showing nodes accounting for 43.31s, 56.16% of 77.12s total
Dropped 256 nodes (cum <= 0.39s)
Showing top 10 nodes out of 126
      flat  flat%   sum%        cum   cum%
    10.11s 13.11% 13.11%     24.15s 31.31%  github.com/ligato/vpp-agent/plugins/kvscheduler/internal/graph.(*edgeLookup).iterSources
     6.42s  8.32% 21.43%      6.42s  8.32%  memeqbody
     5.92s  7.68% 29.11%     11.31s 14.67%  runtime.scanobject
     4.49s  5.82% 34.93%     13.19s 17.10%  runtime.mallocgc
     3.88s  5.03% 39.96%      4.49s  5.82%  runtime.heapBitsSetType
     3.57s  4.63% 44.59%      4.56s  5.91%  runtime.findObject
     3.15s  4.08% 48.68%     10.67s 13.84%  strings.HasPrefix (inline)
     2.39s  3.10% 51.78%      2.39s  3.10%  cmpbody
     1.93s  2.50% 54.28%      1.93s  2.50%  runtime.memequal
     1.45s  1.88% 56.16%      1.45s  1.88%  runtime.duffcopy
```
Now:
```
(pprof) top
Showing nodes accounting for 32.78s, 47.21% of 69.44s total
Dropped 268 nodes (cum <= 0.35s)
Showing top 10 nodes out of 135
      flat  flat%   sum%        cum   cum%
     6.49s  9.35%  9.35%     12.51s 18.02%  runtime.scanobject
     4.76s  6.85% 16.20%     14.08s 20.28%  runtime.mallocgc
     4.72s  6.80% 23.00%      5.95s  8.57%  runtime.findObject
     4.16s  5.99% 28.99%      4.16s  5.99%  runtime.memmove
     3.90s  5.62% 34.61%      4.67s  6.73%  runtime.heapBitsSetType
     2.89s  4.16% 38.77%      2.89s  4.16%  cmpbody
     1.81s  2.61% 41.37%      1.81s  2.61%  runtime.duffcopy
     1.64s  2.36% 43.74%      9.35s 13.46%  sort.Search
     1.21s  1.74% 45.48%      3.81s  5.49%  runtime.wbBufFlush1
     1.20s  1.73% 47.21%      3.39s  4.88%  github.com/ligato/vpp-agent/plugins/kvscheduler/internal/graph.edge.compare
```

The graph is now asymptotically as optimized as it is probably possible. Next optimizations should target limiting the memory usage and perhaps adding some paralelizm into the kvscheduler - best option (= gain / amount-of-work) is to execute descriptor Create/Delete/Update ops in parallel (asynchronously), but leave graph walk sequential.

Signed-off-by: Milan Lenco <milan.lenco@pantheon.tech>